Import Geant4 11.1.0 source tree

+++ source/visualization/Vtk/src/vtkTensorGlyphColor.cxx
@@ -244,7 +244,7 @@ int vtkTensorGlyphColor::RequestData(
     {
       cell = this->GetSource()->GetCell(cellId);
       cellPts = cell->GetPointIds();
-      npts = cellPts->GetNumberOfIds();
+      npts = (int)cellPts->GetNumberOfIds();
       for (dir=0; dir < numDirs; dir++)
       {
         // This variable may be removed, but that
